Move the persistence store interface under AWS KMS to a pkg/store.
- In this way both AWS KMS and IBM Key Protect can leverage this Store interface.
